refactor: directly import typescript methods where they are used
