+++ packages/cargo_tools_vscode/src/logger.rs
@@ -1,8 +1,16 @@
 use tracing::{Event, Level, Subscriber};
 use tracing_subscriber::Layer;
 use tracing_subscriber::layer::Context;
+use wasm_bindgen::prelude::wasm_bindgen;
 
-use crate::vs_code_api::{log_debug, log_error, log_info, log_trace, log_warn};
+#[wasm_bindgen(raw_module = "../runtime.ts")]
+extern "C" {
+    fn log_debug(msg: &str);
+    fn log_info(msg: &str);
+    fn log_warn(msg: &str);
+    fn log_error(msg: &str);
+    fn log_trace(msg: &str);
+}
 
 pub struct VSCodeLogger;
